// client/src/app/interfaces/TrackFeatures.ts
export interface TrackFeatures {
    key: number,                // int
    mode: number,               // int
    time_signature: number,     // int
    acousticness: number,       // float
    danceability: number,       // float
    energy: number,             // float
    instrumentalness: number,   // float
    liveness: number,           // float
    loudness: number,           // float
    speechiness: number,        // float
    valence: number,            // float
    tempo: number,              // float
    popularity: number          // int
}